+++ _pytest/test_synonyms.py
@@ -1,5 +1,5 @@
 from rasa_nlu.extractors.entity_synonyms import EntitySynonymMapper
-from rasa_nlu.pipeline import Interpreter
+from rasa_nlu.model import Interpreter
 
 
 def test_entity_synonyms():
